+++ model_bakery/baker.py
@@ -53,6 +53,7 @@ def make(
     _save_kwargs: Optional[Dict] = None,
     _refresh_after_create: bool = False,
     _create_files: bool = False,
+    _using: str = "",
     **attrs: Any
 ):
     """Create a persisted instance from a given model its associated models.
@@ -61,7 +62,9 @@ def make(
     fields you want to define its values by yourself.
     """
     _save_kwargs = _save_kwargs or {}
-    baker = Baker.create(_model, make_m2m=make_m2m, create_files=_create_files)
+    baker = Baker.create(
+        _model, make_m2m=make_m2m, create_files=_create_files, _using=_using
+    )
     if _valid_quantity(_quantity):
         raise InvalidQuantityException
 
@@ -80,14 +83,18 @@ def make(
 
 
 def prepare(
-    _model: Union[str, Type[ModelBase]], _quantity=None, _save_related=False, **attrs
+    _model: Union[str, Type[ModelBase]],
+    _quantity=None,
+    _save_related=False,
+    _using="",
+    **attrs
 ) -> Model:
     """Create but do not persist an instance from a given model.
 
     It fill the fields with random values or you can specify which
     fields you want to define its values by yourself.
     """
-    baker = Baker.create(_model)
+    baker = Baker.create(_model, _using=_using)
     if _valid_quantity(_quantity):
         raise InvalidQuantityException
 
@@ -105,13 +112,17 @@ def _recipe(name: str) -> Any:
     return import_from_str(".".join((app, "baker_recipes", recipe_name)))
 
 
-def make_recipe(baker_recipe_name, _quantity=None, **new_attrs):
-    return _recipe(baker_recipe_name).make(_quantity=_quantity, **new_attrs)
+def make_recipe(baker_recipe_name, _quantity=None, _using="", **new_attrs):
+    return _recipe(baker_recipe_name).make(
+        _quantity=_quantity, _using=_using, **new_attrs
+    )
 
 
-def prepare_recipe(baker_recipe_name, _quantity=None, _save_related=False, **new_attrs):
+def prepare_recipe(
+    baker_recipe_name, _quantity=None, _save_related=False, _using="", **new_attrs
+):
     return _recipe(baker_recipe_name).prepare(
-        _quantity=_quantity, _save_related=_save_related, **new_attrs
+        _quantity=_quantity, _save_related=_save_related, _using=_using, **new_attrs
     )
 
 
@@ -235,16 +246,18 @@ class Baker(object):
         _model: Union[str, Type[ModelBase]],
         make_m2m: bool = False,
         create_files: bool = False,
+        _using: str = "",
     ) -> "Baker":
         """Create the baker class defined by the `BAKER_CUSTOM_CLASS` setting."""
         baker_class = _custom_baker_class() or cls
-        return baker_class(_model, make_m2m, create_files)
+        return baker_class(_model, make_m2m, create_files, _using=_using)
 
     def __init__(
         self,
         _model: Union[str, Type[ModelBase]],
         make_m2m: bool = False,
         create_files: bool = False,
+        _using: str = "",
     ) -> None:
         self.make_m2m = make_m2m
         self.create_files = create_files
@@ -253,6 +266,7 @@ class Baker(object):
         self.model_attrs = {}  # type: Dict[str, Any]
         self.rel_attrs = {}  # type: Dict[str, Any]
         self.rel_fields = []  # type: List[str]
+        self._using = _using
 
         if isinstance(_model, str):
             self.model = self.finder.get_model(_model)
@@ -309,6 +323,8 @@ class Baker(object):
         **attrs: Any
     ) -> Model:
         _save_kwargs = _save_kwargs or {}
+        if self._using:
+            _save_kwargs["using"] = self._using
 
         self._clean_attrs(attrs)
         for field in self.get_fields():
@@ -341,8 +357,8 @@ class Baker(object):
         instance = self.instance(
             self.model_attrs,
             _commit=commit,
-            _save_kwargs=_save_kwargs,
             _from_manager=_from_manager,
+            _save_kwargs=_save_kwargs,
         )
         if commit:
             for related in self.get_related():
@@ -491,7 +507,7 @@ class Baker(object):
                         m2m_relation.source_field_name: instance,
                         m2m_relation.target_field_name: value,
                     }
-                    make(through_model, **base_kwargs)
+                    make(through_model, _using=self._using, **base_kwargs)
 
     def _remote_field(
         self, field: Union[ForeignKey, OneToOneField]
@@ -534,6 +550,7 @@ class Baker(object):
 
         # attributes like max_length, decimal_places are taken into account when
         # generating the value.
+        field._using = self._using
         generator_attrs = get_required_values(generator, field)
 
         if field.name in self.rel_fields: